+++ CHANGELOG.md
@@ -4,6 +4,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.1.6] - 2023-09-20
+
+### Added
+- Added Vital view for individual person
+- Add measurement modal for Vital view autopopulates vital
+- Person can pan vital graph
+
+### Changed
+
+### Fixed
+
+
 ## [0.1.5] - 2023-09-19
 
 ### Added